requirements.txt
filetype==1.2.0
pillow==10.3.0